+++ algebird-core/src/main/scala/com/twitter/algebird/BloomFilter.scala
@@ -156,7 +156,7 @@ case class BFSparse(hashes : BFHash, bits : CBitSet, width : Int) extends BF {
 
   lazy val numHashes: Int = hashes.size
 
-  lazy val dense : BFInstance = BFInstance(hashes, bits.toBitSet(width), width)
+  @transient lazy val dense : BFInstance = BFInstance(hashes, bits.toBitSet(width), width)
 
   def ++ (other: BF): BF = {
     require(this.width == other.width)
@@ -338,3 +338,13 @@ case class BFHash(numHashes: Int, width: Int, seed: Long = 0L) extends Function1
   }
 }
 
+case class BloomFilterAggregator(bfMonoid: BloomFilterMonoid)  extends MonoidAggregator[String, BF, BF] {
+  val monoid = bfMonoid
+
+  def prepare(value: String) = monoid.create(value)
+  def present(bf: BF) = bf
+}
+
+object BloomFilterAggregator {
+  def apply(numHashes: Int, width: Int, seed: Int = 0): BloomFilterAggregator = BloomFilterAggregator(BloomFilterMonoid(numHashes, width, seed))
+}