+++ algebird-core/src/main/scala/com/twitter/algebird/Group.scala
@@ -32,6 +32,9 @@ trait Group[@specialized(Int,Long,Float,Double) T] extends Monoid[T] {
   def minus(l : T, r : T) : T = plus(l, negate(r))
 }
 
+// For Java interop so they get the default methods
+abstract class AbstractGroup[T] extends Group[T]
+
 // Trivial group. Returns constant on any interaction.
 // The contract is that T be a singleton type (that is, t1 == t2 returns true
 // for all instances t1,t2 of type T).
@@ -82,4 +85,5 @@ object Group extends GeneratedGroupImplicits with ProductGroups {
   implicit val jdoubleGroup : Group[JDouble] = JDoubleField
   implicit def indexedSeqGroup[T:Group]: Group[IndexedSeq[T]] = new IndexedSeqGroup[T]
   implicit def mapGroup[K,V](implicit group : Group[V]) = new MapGroup[K,V]()(group)
+  implicit def scMapGroup[K,V](implicit group : Group[V]) = new ScMapGroup[K,V]()(group)
 }