+++ algebird-core/src/main/scala/com/twitter/algebird/OrderedSemigroup.scala
@@ -28,6 +28,8 @@ object Max {
   def monoid[T](zero: => T)(implicit ord: Ordering[T]): Monoid[Max[T]] =
      Monoid.from(Max(zero)) { (l,r) => if(ord.gteq(l.get, r.get)) l else r }
 
+  def aggregator[T](implicit ord:Ordering[T]): MaxAggregator[T] = MaxAggregator()(ord)
+
   implicit def intMonoid: Monoid[Max[Int]] = monoid(Int.MinValue)
   implicit def longMonoid: Monoid[Max[Long]] = monoid(Long.MinValue)
   implicit def doubleMonoid: Monoid[Max[Double]] = monoid(Double.MinValue)
@@ -61,6 +63,8 @@ object Min {
   def monoid[T](zero: => T)(implicit ord: Ordering[T]): Monoid[Min[T]] =
      Monoid.from(Min(zero)) { (l,r) => if(ord.lteq(l.get, r.get)) l else r }
 
+  def aggregator[T](implicit ord:Ordering[T]): MinAggregator[T] = MinAggregator()(ord)
+
   implicit def intMonoid: Monoid[Min[Int]] = monoid(Int.MaxValue)
   implicit def longMonoid: Monoid[Min[Long]] = monoid(Long.MaxValue)
   implicit def doubleMonoid: Monoid[Min[Double]] = monoid(Double.MaxValue)
@@ -77,4 +81,16 @@ object First {
 case class Last[@specialized(Int,Long,Float,Double) +T](get: T)
 object Last {
   implicit def semigroup[T] = Semigroup.from[Last[T]] { (l,r) => r }
+}
+
+case class MinAggregator[T](implicit ord: Ordering[T]) extends Aggregator[T, T, T] {
+  def prepare(v: T) = v
+  def reduce(l: T, r: T) = ord.min(l , r)
+  def present(v: T) = v
+}
+
+case class MaxAggregator[T](implicit ord: Ordering[T]) extends Aggregator[T, T, T] {
+  def prepare(v: T) = v
+  def reduce(l: T, r: T) = ord.max(l , r)
+  def present(v: T) = v
 }